+++ src/states/controlState.ts
@@ -4,7 +4,7 @@ import { Extent as OlExtent } from 'ol/extent';
 
 import { Time, TimeRange } from '../model/timeSeries';
 import { loadUserSettings } from './userSettings';
-import { getDefaultApiServer } from '../config';
+import { getBranding, getDefaultApiServer } from '../config';
 
 
 export type TimeAnimationInterval = 250 | 500 | 1000 | 2500;
@@ -34,10 +34,12 @@ export interface ControlState {
     legalAgreementAccepted: boolean;
     mapInteraction: MapInteraction;
     imageSmoothingEnabled: boolean;
+    baseMapUrl: string;
 }
 
 
 export function newControlState(): ControlState {
+    const branding = getBranding();
     const state: ControlState = {
         selectedDatasetId: 'local',
         selectedVariableName: 'conc_chl',
@@ -60,6 +62,7 @@ export function newControlState(): ControlState {
         legalAgreementAccepted: false,
         mapInteraction: 'Point',
         imageSmoothingEnabled: false,
+        baseMapUrl: branding.baseMapUrl || 'http://a.tile.osm.org/{z}/{x}/{y}.png',
     };
     return loadUserSettings(state);
 }