+++ CHANGES.md
@@ -1,5 +1,9 @@
 ## Changes in version 0.4.0 (in development)
 
+### Enhancements
+
+* It is now possible to change the base map layer from settings. (#30)
+
 * Different deployments can now be configured through `.env.local` file. (#24)
 
 ## Changes in version 0.3.1